backend/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.3
whitenoise==6.2.0
